setup trait for running kernel
