+++ crates/repl/Cargo.toml
@@ -25,7 +25,7 @@ feature_flags.workspace = true
 futures.workspace = true
 gpui.workspace = true
 image.workspace = true
-jupyter_websocket_client.workspace = true
+jupyter-websocket-client.workspace = true
 language.workspace = true
 log.workspace = true
 markdown_preview.workspace = true
